Refactored to pass in site location to the rendering function and then make the call to query for the time zone. This means bringing in the axios dependency to graph-server to make a service call to the OGCAPI to get the location.
