+++ assets/src/scripts/components/hydrograph/scales.js
@@ -1,12 +1,14 @@
-import { scaleLinear, scaleSymlog } from 'd3-scale';
+import {scaleLinear, scaleSymlog} from 'd3-scale';
 import memoize from 'fast-memoize';
-import { createSelector } from 'reselect';
-import { getYDomain, SYMLOG_PARMS } from './domain';
-import { getLayout } from './layout';
-import { getTimeSeriesForTsKey, TEMPERATURE_PARAMETERS } from './time-series';
-import { visiblePointsSelector, pointsByTsKeySelector } from './drawing-data';
-import { getVariables, getCurrentParmCd, getRequestTimeRange } from '../../selectors/time-series-selector';
-import { convertCelsiusToFahrenheit, convertFahrenheitToCelsius } from '../../utils';
+import {createSelector} from 'reselect';
+
+import {getVariables, getCurrentParmCd, getRequestTimeRange, getTimeSeriesForTsKey} from '../../selectors/time-series-selector';
+import {convertCelsiusToFahrenheit, convertFahrenheitToCelsius} from '../../utils';
+
+import {getYDomain, SYMLOG_PARMS} from './domain';
+import {visiblePointsSelector, pointsByTsKeySelector} from './drawing-data';
+import {getLayout} from './layout';
+import {TEMPERATURE_PARAMETERS} from './time-series';
 
 const REVERSE_AXIS_PARMS = [
     '72019',
