+++ assets/src/scripts/selectors/time-series-selector.js
@@ -2,6 +2,8 @@ import { DateTime } from 'luxon';
 
 import memoize from 'fast-memoize';
 import { createSelector } from 'reselect';
+import uniq from 'lodash/uniq';
+import _includes from 'lodash/includes';
 
 /*
  * Selectors that return properties from the state
@@ -205,3 +207,71 @@ export const isLoadingTS = memoize((tsKey, period, parmCd) => createSelector(
     getTsRequestKey(tsKey, period, parmCd),
     (loadingTSKeys, tsRequestKey) => loadingTSKeys.includes(tsRequestKey)
 ));
+
+/**
+ * Returns a selector that, for a given tsKey and period:
+ * Selects all time series data. If period is null then, the currentDateRange is used
+ * @param  {String} tsKey   Time-series key
+ * @param {String} period or null date range using ISO-8601 duration;
+ * @param  {Object} state   Redux state
+ * @return {Object} - Keys are tsID, values are time-series data
+ */
+export const getTimeSeriesForTsKey = memoize((tsKey, period) => createSelector(
+    getTsRequestKey(tsKey, period),
+    getTimeSeries,
+    (tsRequestKey, timeSeries) => {
+        let x = {};
+        Object.keys(timeSeries).forEach(key => {
+            const series = timeSeries[key];
+            if (series.tsKey === tsRequestKey) {
+                x[key] = series;
+            }
+        });
+        return x;
+    }
+));
+
+
+/**
+ * Returns a selector that, for a given tsKey:
+ * Selects all time series for the current time series variable and current date range.
+ * @param  {String} tsKey   Time-series key
+ * @param  {Object} state   Redux state
+ * @return {Object}         Time-series data
+ */
+export const getCurrentVariableTimeSeries = memoize((tsKey, period) => createSelector(
+    getTimeSeriesForTsKey(tsKey, period),
+    getCurrentVariable,
+    (timeSeries, variable) => {
+        let ts = {};
+        if (variable) {
+            Object.keys(timeSeries).forEach(key => {
+                const series = timeSeries[key];
+                if (series.variable === variable.oid) {
+                    ts[key] = series;
+                }
+            });
+        }
+        return ts;
+    }
+));
+
+/*
+ * @param {String} tsKey - current or compare
+ * @param {String} or null period - date range of interest specified as an ISO-8601 duration. If null, P7D is assumed
+ * @param {String} or null parmCd - Only need to specify if period is something other than P7D or null
+ * return {Array} an object containing method_id and method_description properties
+ */
+export const getAllMethodsForCurrentVariable = createSelector(
+    getMethods,
+    getCurrentVariableTimeSeries('current', 'P7D'),
+    (methods, timeSeries) => {
+        const allMethods = Object.values(methods);
+        const currentMethodIds = uniq(Object.values(timeSeries).map((ts) => ts.method));
+
+        return allMethods.filter((method) => {
+            return _includes(currentMethodIds, method.methodID);
+        });
+    }
+);
+
