+++ assets/tests/scripts/global-config.js
@@ -1,11 +1,7 @@
 var CONFIG = {
     TIMESERIES_AUDIO_ENABLED: true,
     MULTIPLE_TIME_SERIES_METADATA_SELECTOR_ENABLED: true,
-    FIM_GIS_ENDPOINT: 'https:/fakelegendservice.com',
-    siteLocation: {
-        latitude: 43.5,
-        longitude: -100.0
-    }
+    FIM_GIS_ENDPOINT: 'https:/fakelegendservice.com'
 };
 
 var mockOscillator = function() {
