+++ graph-server/package.json
@@ -22,6 +22,7 @@
   },
   "homepage": "https://github.com/usgs/waterdataui#readme",
   "dependencies": {
+    "axios": "^0.19.2",
     "body-parser": "1.19.0",
     "express": "4.17.1",
     "express-cache-headers": "0.1.4",
