+++ wdfn-server/waterdata/templates/macros/components.html
@@ -1,5 +1,5 @@
-{% macro TimeSeriesComponent(site_no) -%}
-    <div class="wdfn-component" data-component="hydrograph" data-siteno="{{ site_no }}">
+{% macro TimeSeriesComponent(site_no, latitude, longitude) -%}
+    <div class="wdfn-component" data-component="hydrograph" data-siteno="{{ site_no }}" data-latitude="{{ latitude }}" data-longitude="{{ longitude }}">
         <div class="loading-indicator-container"></div>
         <div class="graph-container"></div>
         <div class="select-time-series-container"></div>
