+++ wdfn-server/waterdata/templates/monitoring_location.html
@@ -28,14 +28,6 @@
         {{ json_ld }}
     </script>
     {% endif %}
-    {% if stations|length == 1 %}
-        <script type="text/javascript">
-            CONFIG['siteLocation'] = {
-                latitude: {{ stations[0].dec_lat_va }},
-                longitude: {{ stations[0].dec_long_va }}
-            };
-        </script>
-    {% endif %}
 {% endblock page_script %}
 
 {% set body_id = 'monitoring-location' %}
@@ -77,7 +69,7 @@
                     <p id="site-description">{{ components.Description(stations[0].site_no, location_with_values, parm_grp_summary) }}</p>
                 </div>
 
-                {{ components.TimeSeriesComponent(stations[0].site_no) }}
+                {{ components.TimeSeriesComponent(stations[0].site_no, stations[0].dec_lat_va, stations[0].dec_long_va) }}
 
                 {% if monitoring_camera.exists %}
                     <div class="usa-accordion" style="margin-top:1em;">
